pyproject.toml
[project]
name = "paper-summary-automation"
version = "0.1.0"
description = "Automate downloading a research paper PDF, summarizing it with OpenAI, and posting the result to Scrapbox."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.30.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
paper-summary = "paper_summary:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["paper_summary.py"]
